requirements.txt
requests
pandas
scikit-learn
seaborn
matplotlib
streamlit
sqlalchemy
mysql-connector-python
python-dotenv
